[DEPLOY] appsepc.yml 수정

+++ scripts/start-server.sh
@@ -9,6 +9,11 @@ echo "현재 디렉토리: $(pwd)"
 echo "파일 목록:"
 ls -al
 
+if [ ! -f docker-compose.yml ]; then
+  echo "❌ docker-compose.yml 파일이 존재하지 않습니다."
+  exit 1
+fi
+
 docker compose down || true
 docker compose up -d
